requirements.txt
scipy
scikit-learn
numpy
Pillow
numpyencoder
matplotlib
tqdm
ftfy
regex
opencv-python
blobfile
timm
transformers
fairscale